Move helpers/setup into partial
